Add new funding feed view
