+++ src/feed/views/__init__.py
@@ -1,0 +1,11 @@
+"""
+This module contains all the feed-related views.
+"""
+
+from .feed_view import FeedViewSet
+from .funding_feed_view import FundingFeedViewSet
+
+__all__ = [
+    "FeedViewSet",
+    "FundingFeedViewSet",
+]
